+++ GraphiQL-App/src/pages/WelcomePage/WelcomePage.tsx
@@ -2,7 +2,6 @@ import { NavLink } from 'react-router-dom';
 import { useAuthState } from 'react-firebase-hooks/auth';
 import { db, auth } from '../../helpers/firebase';
 import { query, collection, getDocs, where } from 'firebase/firestore';
-
 import { useEffect } from 'react';
 import { useAppDispatch, useAppSelector } from '../../store/hooksRedux';
 import { exitUser, setName, UserState } from '../../store/slices/userSlice';
