+++ GraphiQL-App/src/store/reduxStore.tsx
@@ -4,7 +4,6 @@ import userReducer from './slices/userSlice';
 import docPaneStateSlice from './slices/docPaneStateSlice';
 import queryPanelStateSlice from './slices/queryPanelStateSlice';
 import queryParametersSlice from './slices/queryParametersSlice';
-
 import langSlice from './slices/langSlice';
 
 const rootReducer = combineReducers({
